1.0.1

+++ src/utils/FileChooser.java
@@ -30,6 +30,13 @@ public class FileChooser {
         if (i == jFileChooser.APPROVE_OPTION) {
             fileDirPath = jFileChooser.getSelectedFile().getAbsolutePath();
         }
+        //判断用户选取的路径文件夹中是否为空
+        if (new File(fileDirPath).length() != 0) {
+            int comfirmInt = JOptionPane.showConfirmDialog(null, "所选文件夹不为空，是否继续");
+            if (comfirmInt != JOptionPane.YES_OPTION) {
+                return;
+            }
+        }
         BufferedWriter bw = null;
         try {
             bw = new BufferedWriter(new FileWriter(file));
